feat(security): ai.ops(long) overload pour prévenir les overflows int (#3652)
Suite au fix overflow OOM dans replace() (#5769754), les reviews ont
convergé sur une fix structurelle plutôt qu'un point patch.

Ajout de AI.ops(long) qui sature à Integer.MAX_VALUE et délègue à
ops(int). Permet aux call sites d'écrire :

    ai.ops((long) string.length() * replace.length())

sans risque d'overflow ni saturation manuelle.

- StringClass.replace : retire la saturation locale, utilise ops(long).
- JSONClass.jsonEncode/jsonDecode : widen json.length() * 10 vers (long)
  par defense in depth (overflow théorique à 215M chars, low
  exploitability mais le fix est trivial avec le nouvel overload).

Les autres call sites avec multiplication (MapLeekValue, SetLeekValue,
ArrayLeekValue avec 5*size*log(size)) sont bornés par RAM cap en
pratique. À widen progressivement dans un sweep dédié si besoin.

Issue #3652

+++ src/main/java/leekscript/runner/classes/JSONClass.java
@@ -13,7 +13,7 @@ public class JSONClass {
 		try {
 
 			String json = Json.toJson(ai.toJSON(value));
-			ai.ops(json.length() * 10);
+			ai.ops((long) json.length() * 10);
 			return json;
 
 		} catch (Exception e) {
@@ -31,7 +31,7 @@ public class JSONClass {
 		try {
 
 			var obj = LeekValueManager.parseJSON(Json.parse(json), ai);
-			ai.ops(json.length() * 10);
+			ai.ops((long) json.length() * 10);
 			return obj;
 
 		} catch (Throwable e) {
